修改进会启动逻辑start->main

+++ app/src/main/java/org/dync/teameeting/ui/activity/MainActivity.java
@@ -150,6 +150,7 @@ public class MainActivity extends BaseActivity {
     protected void onCreate(Bundle savedInstanceState) {
 
         super.onCreate(savedInstanceState);
+        TeamMeetingApp.isInitFalg = true;
         mContext = this;
         setContentView(R.layout.activity_main);
         TeamMeetingApp.setMainActivity(this);
@@ -159,6 +160,36 @@ public class MainActivity extends BaseActivity {
             Log.e(TAG, "onCreate: " + TeamMeetingApp.getmSelfData().getMeetingLists().toString());
         }
         updataApp();
+    }
+
+
+    /**
+     * 再次启动
+     *
+     * @param intent
+     */
+    @Override
+    protected void onNewIntent(Intent intent) {
+        super.onNewIntent(intent);
+        String urlMeetingId = intent.getStringExtra("urlMeetingId");
+        if (mDebug)
+            Log.e(TAG, "onNewIntent-----: urlMeetingId" + urlMeetingId);
+        if (urlMeetingId != null) {
+            upDataMeetingList();
+            int position = MeetingHelper.getMeetingIdPosition(mRoomMeetingList, urlMeetingId);
+            if (position >= 0) {
+                meetingPositiotrue(position);
+            } else {
+                Toast.makeText(mContext, R.string.str_join_room_wait, Toast.LENGTH_LONG);
+                mNetWork.getMeetingInfo(urlMeetingId, JoinActType.JOIN_LINK_JOIN_ACTIVITY);
+            }
+        }
+    }
+
+    @Override
+    protected void onRestart() {
+        super.onRestart();
+        Log.e(TAG, "onRestart: ");
     }
 
     private void updataApp() {
@@ -188,6 +219,7 @@ public class MainActivity extends BaseActivity {
 
             @Override
             public void onNoUpdateAvailable() {
+
             }
         });
 
@@ -196,20 +228,30 @@ public class MainActivity extends BaseActivity {
 
     private void initdata() {
         mWarningCancel = DialogHelper.createWarningCancel(mContext);
+
         upDataMeetingList();
+
         mMsgSender = TeamMeetingApp.getmMsgSender();
         isSetUserName = LocalUserInfo.getInstance(mContext).getUserInfoBoolean(LocalUserInfo.SET_USER_NAME);
         Intent intent = getIntent();
         isNotifactionChack = intent.getBooleanExtra("isNotifactionChack", false);
         mUrlMeetingId = intent.getStringExtra("urlMeetingId");
 
+        urlMeetingIdEnterRoom(intent);
+    }
+
+    /**
+     * url enter room　判断
+     *
+     * @param intent
+     */
+    private void urlMeetingIdEnterRoom(Intent intent) {
         if (isNotifactionChack) {
             int position = MeetingHelper.getMeetingIdPosition(mRoomMeetingList, mUrlMeetingId);
             mNotifTags = intent.getIntExtra("tags", 0);
-            if (position > 0) {
+            if (position > -1) {
                 enterMeetingActivity(position);
             }
-
             return;
         }
 
@@ -220,9 +262,7 @@ public class MainActivity extends BaseActivity {
             }
             int position = TeamMeetingApp.getmSelfData().getMeetingIdPosition(mUrlMeetingId);
             if (position >= 0) {
-
                 meetingPositiotrue(position);
-
             } else {
                 Toast.makeText(mContext, R.string.str_join_room_wait, Toast.LENGTH_LONG);
                 mNetWork.getMeetingInfo(mUrlMeetingId, JoinActType.JOIN_LINK_JOIN_ACTIVITY);
@@ -352,7 +392,7 @@ public class MainActivity extends BaseActivity {
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        mNetWork.signOut(mSign);
+
         boolean pushStopped = JPushInterface.isPushStopped(this);
         if (mDebug)
             Log.e(TAG, "onDestroy=--- !+pushStopped" + pushStopped);
@@ -697,10 +737,8 @@ public class MainActivity extends BaseActivity {
                 mExitTime = System.currentTimeMillis();
             } else {
                 mSign = getSign();
-
-                 Intent home = new Intent(Intent.ACTION_MAIN);
-                 home.addCategory(Intent.CATEGORY_HOME);
-                 startActivity(home);
+                mNetWork.signOut(mSign);
+                this.finish();
             }
             return true;
         }
@@ -735,6 +773,8 @@ public class MainActivity extends BaseActivity {
                 settingReName(data);
                 break;
             case ExtraType.RESULT_CODE_ROOM_SETTING_DELETE:
+                if (mDebug)
+                    Log.e(TAG, "RESULT_CODE_ROOM_SETTING_DELETE: 关闭");
                 seetingDeleteRoom(data);
                 break;
             case ExtraType.RESULT_CODE_ROOM_SETTING_CLOSE:
@@ -773,6 +813,8 @@ public class MainActivity extends BaseActivity {
     }
 
     private void seetingDeleteRoom(Intent data) {
+        if (mDebug)
+            Log.e(TAG, "seetingDeleteRoom: ");
         mSign = getSign();
         int position = data.getIntExtra("position", 0);
         String meetingId = data.getStringExtra("meetingId");
@@ -799,13 +841,16 @@ public class MainActivity extends BaseActivity {
         }
     }
 
+    /**
+     * init roomMeetingList
+     */
     private void upDataMeetingList() {
         List<MeetingListEntity> list = TeamMeetingApp.getmSelfData().getMeetingLists();
         for (int i = 0; i < list.size(); i++) {
             list.get(i).initUnReadMessage(mContext);
         }
         if (list != null) {
-            mRoomMeetingList.clear();
+            //mRoomMeetingList.clear();
             //mRoomMeetingList.addAll();
             mRoomMeetingList = list;
             Logger.e(list.toString() + "----" + mRoomMeetingList.toString());
@@ -855,7 +900,9 @@ public class MainActivity extends BaseActivity {
         String joinType;
         switch (usable) {
             case 0://no
+                //// TODO: 2016/3/22  会议已经被删除， 非房主应该删除
                 Toast.makeText(mContext, R.string.str_meeting_deleted, Toast.LENGTH_SHORT).show();
+
                 break;
             case 1:
 
@@ -875,7 +922,6 @@ public class MainActivity extends BaseActivity {
                     // statrMeetingActivity(mUrlMeetingName, meetinId, anyrtcId);
                     statrMeetingActivity(meetingListEntity);
                 } else if (joinType == JoinActType.JOIN_LINK_JOIN_ACTIVITY) {
-
                     Toast.makeText(mContext, R.string.str_meeting_privated, Toast.LENGTH_SHORT).show();
                 }
 
@@ -1018,16 +1064,18 @@ public class MainActivity extends BaseActivity {
             case MSG_GET_MEETING_INFO_SUCCESS:
                 if (mDebug)
                     Log.e(TAG, "MSG_GET_MEETING_INFO_SUCCESS");
-                if (msg.getData().getString(JoinActType.JOIN_TYPE).equals(JoinActType.JOIN_ENTER_ACTIVITY)) {
-                    getMeetingInfoSuccess(msg);
-                }
+                if (msg.getData().getString(JoinActType.JOIN_TYPE).equals(JoinActType.JOIN_ENTER_ACTIVITY)
+                        || msg.getData().getString(JoinActType.JOIN_TYPE).equals(JoinActType.JOIN_LINK_JOIN_ACTIVITY)){
+                getMeetingInfoSuccess(msg);
+            }
 
-                break;
+            break;
             case MSG_GET_MEETING_INFO_FAILED:
                 if (mDebug)
                     Log.e(TAG, "MSG_GET_MEETING_INFO_FAILED");
 
-                if (msg.getData().getString(JoinActType.JOIN_TYPE).equals(JoinActType.JOIN_LINK_JOIN_ACTIVITY)) {
+                String join_type = msg.getData().getString(JoinActType.JOIN_TYPE);
+                if (join_type.equals(JoinActType.JOIN_LINK_JOIN_ACTIVITY) || join_type.equals(JoinActType.JOIN_ENTER_ACTIVITY)) {
                     meetingId = msg.getData().getString("meetingid");
                     mNetWork.deleteRoom(getSign(), meetingId);
                     Toast.makeText(mContext, R.string.meeting_delete_create, Toast.LENGTH_SHORT).show();
@@ -1059,11 +1107,16 @@ public class MainActivity extends BaseActivity {
                     Log.e(TAG, "MSG_UP_DATE_USER_MEETING_JOIN_TIME_SUCCESS " + mAdapter.getCount());
                 break;
             case MSG_DELETE_ROOM_SUCCESS:
+                if (mDebug) {
+                    Log.e(TAG, "onEventMainThread: MSG_DELETE_ROOM_SUCCESS");
+                }
                 meetingId = msg.getData().getString("meetingid");
                 int position = mAdapter.getMeetingIdPosition(meetingId);
-                mRoomMeetingList.remove(position);
-                mAdapter.notifyDataSetChanged();
-                CRUDChat.deleteByMeetingId(mContext, meetingId);
+                if (position != -1) {
+                    mRoomMeetingList.remove(position);
+                    mAdapter.notifyDataSetChanged();
+                    CRUDChat.deleteByMeetingId(mContext, meetingId);
+                }
                 break;
             case JOIN_MEETINGID_EXIST:
                 int pos = (int) msg.obj;
@@ -1086,13 +1139,7 @@ public class MainActivity extends BaseActivity {
                     Log.e(TAG, "MSG_NOTIFICATION_MAIN");
                 msgNotificatinMain(msg);
                 break;
-            case MSG_URL_MEETING_EXIT:
-                if (mDebug)
-                    Log.e(TAG, "MSG_URL_MEETING_EXIT");
-                Intent intent = new Intent(MainActivity.this, MainActivity.class);
-                startActivity(intent);
 
-                break;
             case MSG_URL_START_MEETING:
                 if (mDebug)
                     Log.e(TAG, "MSG_URL_START_MEETING: ");
@@ -1113,6 +1160,14 @@ public class MainActivity extends BaseActivity {
                 }
 
                 break;
+
+            case MSG_ROOMSEETING_ENTER_ROOM:
+                if (mDebug)
+                    Log.e(TAG, "onEventMainThread: MSG_ROOMSEETING_ENTER_ROOM");
+                int posation = (int) msg.obj;
+                if (posation >= 0) {
+                    enterMeetingActivity(posation);
+                }
             default:
                 break;
         }
